Add Character System
This commit adds the character system into the game. with this commit comes a structure that will allow us to select our character based on a character prefab. Each character has its unique weapons and abilities assigned to it based on a character scriptable object reference. This holds all the data about what base abilities and weapons the character will have, as well as general character stats for shared character logic.

The only thing this commit does not do is allow the player to select their character on startup.