Characters (#60)
* Add Charged Shot Weapon

This commit adds the charged weapon shot, fully functioning. The weapon is currently equipped on the player and will have to be toggled on/off depending on the selected character.

I aimed for full control over the weapon's functionality. Every field that we would want to tinker with has been serialized. The only thing not configured is the connection to the augment system. Here's a list of lever we can pull:
-fire rate
-damage brackets
-base delay between shots
-random additional delay between shots
-total hit scanners (the things that charge the weapon)
-base max damage bonus
-perfect accuracy damage bonus

* Update Charged Weapon Functionality

This commit updates and simplifies the charged weapon with simpler code. Also, this commit adds some better control over the elements on screen.

Fun fact: you can store coroutines in a variable and stop that specific coroutine by the variable reference

* Add Shotgun W/ Shotgun Spread

This commit adds the final weapon into our game: the shotgun. The shotgun has a spread value that can be controller, though high spread values may result in unwanted results. Try keeping the spread beneath 5 degrees for accurate results.

There were issues with the UI crosshair. I tried to make it shrink/expand in regards to the spread of the shotgun, but I could not get an accurate crosshair

* Move Scripts to Correct Folder

* Add Character System

This commit adds the character system into the game. with this commit comes a structure that will allow us to select our character based on a character prefab. Each character has its unique weapons and abilities assigned to it based on a character scriptable object reference. This holds all the data about what base abilities and weapons the character will have, as well as general character stats for shared character logic.

The only thing this commit does not do is allow the player to select their character on startup.